FrostEffectController.cs 作成。CameraController.cs 内の FrostEffect 関連の処理を移植して役割分担。 各スクリプトのリファクタリング。コメントの添削。FrostEffectObject.cs の名称を FrostEffector.cs に変更。
