+++ Assets/Prefabs/SnowyTree.prefab
@@ -58,6 +58,7 @@ GameObject:
   - component: {fileID: 3001741863889458178}
   - component: {fileID: 2974775407665491018}
   - component: {fileID: 2974775407665491019}
+  - component: {fileID: 1355123874}
   m_Layer: 0
   m_Name: SnowyTree
   m_TagString: Untagged
@@ -72,7 +73,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2974775407665491016}
-  m_LocalRotation: {x: 0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
@@ -174,3 +175,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   duration: 3
+--- !u!114 &1355123874
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2974775407665491016}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9c957f081e4443b49a61d80977dd940d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
