+++ Assets/Scripts/CameraController.cs
@@ -9,17 +9,9 @@ public class CameraController : MonoBehaviour
 
     private Vector3 offset;
 
-    [SerializeField]
-    private FrostEffect frostEffect;
-
-    private float currentFrostValue;
-
     void Start() {
         // カメラと追従対象のゲームオブジェクトとの距離を補正値として取得
         offset = transform.position - playerObj.transform.position;
-
-        // 霜エフェクトの初期化
-        InitialFrostAmount();
     }
 
 
@@ -30,38 +22,5 @@ public class CameraController : MonoBehaviour
             // カメラの位置を追従対象の位置 + 補正値にする
             transform.position = playerObj.transform.position + offset;
         }
-
-        if (Input.GetKeyDown(KeyCode.Z)) {
-            UpdateFrostAmount(0.1f);
-        }
-
-        // 霜でおおわれているとき
-        if (currentFrostValue > 0) {
-
-            // 少しずつ見えるようにする
-            currentFrostValue -= Time.deltaTime / 20;
-            UpdateFrostAmount(0);
-
-            if (currentFrostValue <= 0) {
-                currentFrostValue = 0;
-            }
-        }
-    }
-
-    /// <summary>
-    /// FrostEffectの状態を更新
-    /// </summary>
-    /// <param name="amount"></param>
-    public void UpdateFrostAmount(float amount) {
-        currentFrostValue += amount;
-        frostEffect.FrostAmount = currentFrostValue;
-    }
-
-    /// <summary>
-    /// FrostEffectの初期化
-    /// </summary>
-    public void InitialFrostAmount() {
-        currentFrostValue = 0;
-        frostEffect.FrostAmount = currentFrostValue;
     }
 }
