+++ Assets/Scripts/FrostEffectController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d40dcc82284615a40bc57ab523c305d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
